// src/app/models/user.ts
export interface User {
    nom: string;
    scoreUn: number;
    scoreDeux: number;
    scoreTrois: number;
    scoreQuatre: number;
    scoreCinq: number;
    scoreSix: number;
    scoreBrelan: number;
    scorePSuite: number;
    scoreGSuite: number;
    scoreFull: number;
    scoreCarre: number;
    scoreYam: number;
}